refactor tooltip inline styling
